FederatedLearningCode/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
minio==7.1.1
numpy==1.21.2
pip-chill==1.0.1
PyYAML==5.4.1
requests==2.26.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
torch~=1.9.1